[ESP8266] Fix build issues due to older compiler for ESP8266 2.7.4

+++ src/src/DataStructs/KeyValueStruct.h
@@ -58,12 +58,12 @@ struct KeyValueStruct
 
   KeyValueStruct(const String         & key,
                  const float          & val,
-                 int                    nrDecimals = 4,
+                 uint8_t                    nrDecimals = 4,
                  Format                 format     = Format::Default);
 
   KeyValueStruct(const String         & key,
                  const double         & val,
-                 int                    nrDecimals = 4,
+                 uint8_t                    nrDecimals = 4,
                  Format                 format     = Format::Default);
